separated Smart Cabin Door code from index.ts
